Add tailwindcss-animate config
