+++ lib/generators/ruby_ui/install/templates/tailwind.config.js.js-package.erb
@@ -71,4 +71,7 @@ module.exports = {
       },
     },
   },
+  plugins: [
+    require("tailwindcss-animate"),
+  ],
 }
