+++ lib/generators/ruby_ui/install/templates/tailwind.config.js.tailwindcss-rails.erb
@@ -77,5 +77,6 @@ module.exports = {
     require('@tailwindcss/forms'),
     require('@tailwindcss/typography'),
     require('@tailwindcss/container-queries'),
+    require("../vendor/javascript/tailwindcss-animate")
   ]
 }
